fin_backend/pyproject.toml

[tool.poetry]
name = "fin-backend"
version = "0.1.0"
description = "Enterprise Fintech Backend Architecture"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "core"}, {include = "apps"}, {include = "common"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"
psycopg2-binary = "^2.9.11"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.4.0"}
drf-spectacular = "^0.29.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"      # The modern, lightning-fast Python linter/formatter (replaces Flake8/Black)
pytest = "^8.1.1"
pytest-django = "^4.8.0"
black = "^26.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration (Strict linting for clean code)
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"] # Flake8 errors, Pyflakes, Isort (import sorting)
ignore = []
